rename KeplerWallet to CosmosWallet
