+++ web/src/features/CosmosWallet/utils/utils.test.ts
